Fix undefined names and real bugs found by pyflakes across 24 files
- mesh_factory.py: fix typo Excetion -> Exception
- esri2sww.py: restore missing imports (os, num, log, Geo_reference, netcdf_*, get_min_max_indices)
- sww2pts.py, dem2dem.py, dem2pts.py, grd2array.py, urs2txt.py: fix missing imports and wrong variable names
- print_stats.py: replace removed numpy type aliases Int8/Float with int8/float64
- polygon.py: remove dead NEW_C_intersection (used undefined _intersection C ext, never called)
- urs.py, csv_file.py: add missing ANUGAError and inside_polygon imports; fix Python 2 `file` builtin reference
- friction.py, forcing.py, mannings_operator.py: add missing log import
- elliptic_operator.py, kinematic_viscosity_operator.py: fix __rmul__ bug (new = matrix * new -> matrix * other)
- set_w_uh_vh_operator.py: fix undefined domain -> self.domain, stage -> w_uh_vh[0]
- parallel_structure_operator.py: fix none -> None
- boyd_pipe_operator.py, parallel_boyd_box_operator.py, parallel_weir_orifice_trapezoid_operator.py: fix undefined culvert_type in debug log
- distribute_mesh.py: fix partition_mesh_without_map return (was returning undefined nodes/ttriangles/boundary)
- partitioning.py: remove dead METIS 4 branch using undefined partMeshNodal
- sequential_distribute.py: remove broken debug assertions using undefined new_triangles/new_nodes/original_triangles

+++ anuga/parallel/print_stats.py
@@ -19,7 +19,7 @@ import sys
 
 from anuga.utilities import parallel_abstraction as pypar
 
-from numpy import array, zeros, ones, take, nonzero, float
+from numpy import array, zeros, ones, take, nonzero, int8, float64
 from anuga.utilities.norms import l1_norm, l2_norm, linf_norm
 
 #########################################################
@@ -44,7 +44,7 @@ from anuga.utilities.norms import l1_norm, l2_norm, linf_norm
 
 def build_full_flag(domain, ghost_recv_dict):
 
-    tri_full_flag = ones(len(domain.get_triangles()), Int8)
+    tri_full_flag = ones(len(domain.get_triangles()), int8)
     for i in list(ghost_recv_dict.keys()):
         for id in ghost_recv_dict[i][0]:
             tri_full_flag[id] = 0
@@ -73,8 +73,8 @@ def print_l1_stats(full_edge):
     numprocs = pypar.size()
     myid = pypar.rank()
     
-    tri_norm = zeros(3, Float)
-    recv_norm = zeros(3, Float)
+    tri_norm = zeros(3, float64)
+    recv_norm = zeros(3, float64)
     tri_norm[0] = l1_norm(full_edge[:, 0])
     tri_norm[1] = l1_norm(full_edge[:, 1])
     tri_norm[2] = l1_norm(full_edge[:, 2])
@@ -110,8 +110,8 @@ def print_l2_stats(full_edge):
     numprocs = pypar.size()
     myid = pypar.rank()
     
-    tri_norm = zeros(3, Float)
-    recv_norm = zeros(3, Float)
+    tri_norm = zeros(3, float64)
+    recv_norm = zeros(3, float64)
     tri_norm[0] = pow(l2_norm(full_edge[:, 0]), 2)
     tri_norm[1] = pow(l2_norm(full_edge[:, 1]), 2)
     tri_norm[2] = pow(l2_norm(full_edge[:, 2]), 2)
@@ -148,8 +148,8 @@ def print_linf_stats(full_edge):
     numprocs = pypar.size()
     myid = pypar.rank()
     
-    tri_norm = zeros(3, Float)
-    recv_norm = zeros(3, Float)
+    tri_norm = zeros(3, float64)
+    recv_norm = zeros(3, float64)
         
     tri_norm[0] = linf_norm(full_edge[:, 0])
     tri_norm[1] = linf_norm(full_edge[:, 1])
